+++ src/baobab_scryfall_api_caller/models/cards/__init__.py
@@ -11,6 +11,7 @@ from baobab_scryfall_api_caller.models.cards.card_collection_identifier import (
 from baobab_scryfall_api_caller.models.cards.card_collection_result import CardCollectionResult
 from baobab_scryfall_api_caller.models.cards.card_face import CardFace
 from baobab_scryfall_api_caller.models.cards.card_search_query import CardSearchQuery
+from baobab_scryfall_api_caller.models.cards.image_uris import ImageUris
 
 __all__ = [
     "AutocompleteResult",
@@ -19,5 +20,6 @@ __all__ = [
     "CardCollectionResult",
     "CardFace",
     "CardSearchQuery",
+    "ImageUris",
     "MAX_CARD_COLLECTION_IDENTIFIERS",
 ]